refactor(project-files-view): extract presentation owner
